+++ test/common/http/conn_pool_grid_test.cc
@@ -909,7 +909,7 @@ TEST_F(ConnectivityGridTest, Http3FailedRecentlyThenFailsAgain) {
 } // namespace Envoy
 
 #include "test/mocks/server/transport_socket_factory_context.h"
-#include "source/common/quic/quic_transport_socket_factory.h"
+#include "source/common/quic/quic_client_transport_socket_factory.h"
 namespace Envoy {
 namespace Http {
 namespace {
