+++ test/common/quic/BUILD
@@ -361,7 +361,8 @@ envoy_cc_test(
     ],
     tags = ["nofips"],
     deps = [
-        "//source/common/quic:quic_transport_socket_factory_lib",
+        "//source/common/quic:quic_client_transport_socket_factory_lib",
+        "//source/common/quic:quic_server_transport_socket_factory_lib",
         "//source/extensions/transport_sockets/tls:context_config_lib",
         "//test/mocks/server:transport_socket_factory_context_mocks",
         "//test/mocks/ssl:ssl_mocks",
