+++ test/common/quic/client_connection_factory_impl_test.cc
@@ -1,7 +1,7 @@
 #include <chrono>
 
 #include "source/common/quic/client_connection_factory_impl.h"
-#include "source/common/quic/quic_transport_socket_factory.h"
+#include "source/common/quic/quic_client_transport_socket_factory.h"
 
 #include "test/common/upstream/utility.h"
 #include "test/mocks/common.h"
