+++ test/integration/BUILD
@@ -2371,8 +2371,8 @@ envoy_cc_test(
             "//source/common/quic:active_quic_listener_lib",
             "//source/common/quic:client_connection_factory_lib",
             "//source/common/quic:quic_client_factory_lib",
+            "//source/common/quic:quic_client_transport_socket_factory_lib",
             "//source/common/quic:quic_server_factory_lib",
-            "//source/common/quic:quic_transport_socket_factory_lib",
             "//test/integration/filters:pause_filter_for_quic_lib",
         ],
     }),
