+++ test/integration/listener_extension_discovery_integration_test.cc
@@ -774,7 +774,7 @@ TEST_P(ListenerExtensionDiscoveryIntegrationTest, TwoSubscriptionsConfigDumpWith
 
 #include "quiche/quic/core/deterministic_connection_id_generator.h"
 #include "source/common/quic/client_connection_factory_impl.h"
-#include "source/common/quic/quic_transport_socket_factory.h"
+#include "source/common/quic/quic_server_transport_socket_factory.h"
 #include "test/integration/utility.h"
 
 namespace Envoy {
