+++ test/integration/utility.cc
@@ -26,7 +26,7 @@
 
 #ifdef ENVOY_ENABLE_QUIC
 #include "source/common/quic/client_connection_factory_impl.h"
-#include "source/common/quic/quic_transport_socket_factory.h"
+#include "source/common/quic/quic_client_transport_socket_factory.h"
 #include "quiche/quic/core/deterministic_connection_id_generator.h"
 #endif
