+++ tools/code_format/config.yaml
@@ -142,7 +142,7 @@ paths:
     - source/common/config/datasource.cc
     - source/common/config/utility.cc
     - source/common/runtime/runtime_impl.cc
-    - source/common/quic/quic_transport_socket_factory.cc
+    - source/common/quic/quic_server_transport_socket_factory.cc
     - source/common/filter/config_discovery_impl.cc
     - source/common/json/json_internal.cc
     - source/common/router/scoped_rds.cc
